requirements.txt
streamlit
pandas
sqlalchemy
pillow
